#14 rename
to untie the gordian knot this folder has become for me, i rename it
appropriately. The new `./reports` will take only the tools I want
people to use. @Maleeha , this will invariably affect the scripts
referencing it, and I would love to learn how to undertake a rename like
this using the regexes.