elsa data ready in shape
@wibeasley , the data appears to be ready in shape, but not in content.
There are issues I cannot explain:  (the same value across waves for
fev, pef, fvc ). Besides, in the wide format it shows values for these
variables at the timepoints where there are not supposed to be any.
However, i think we can move on, leaving the data quality for later,
when the syntax  is ready. Starting PCS now